Fix filepath and dont reinitialize yleisviestijson if it already exists

+++ run_yleisviesti.sh
@@ -6,8 +6,10 @@ cd yleisviestipalvelu
 mkdir -p messages/archive
 
 echo "${HTPASSWD}" > /etc/nginx/auth.htpasswd
-echo '{"staticMessages": []}' > messages/yleisviesti.json
+if [ ! -f "messages/yleisviesti.json" ]; then
+    echo '{"staticMessages": []}' > messages/yleisviesti.json
+fi
 
-python main.py --port=${PORT} --filename=messages/yleisviesti.json --filepath=messages/archive &
+python main.py --port=${PORT} --filename=messages/yleisviesti.json --filepath=messages/archive/ &
 
 /usr/sbin/nginx
